% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_est.R
\name{coef.parms_mml}
\alias{coef.parms_mml}
\title{Extract parameters from MML fit object}
\usage{
\method{coef}{parms_mml}(object, what = c("items", "populations"), ...)
}
\arguments{
\item{object}{object returned by fit_1pl or fit_2pl}

\item{what}{parameters to extract}

\item{...}{ignored}
}
\value{
data.frame of parameters
}
\description{
Extract parameters from MML fit object
}
