% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_est.R
\docType{package}
\name{dexterMML-package}
\alias{dexterMML}
\alias{dexterMML-package}
\title{DexterMML: MML addition to dexter}
\description{
For use in cases where CML is not possible, such as an adaptive test
or a multi stage test with unknown routing rules (it happens).
}
\details{
In all other cases, using CML (i.e. dexter or dexterMST) is strongly preferrable.

DexterMML distinguishes itself from other MML R-packages by:

\itemize{
\item including far fewer models and options
\item being considerably faster
\item support for the dexter data(base) structure
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dexter-psychometrics.github.io/dexter/}
  \item Report bugs at \url{https://github.com/dexter-psychometrics/dexterMML/issues}
}

}
