% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_est.R
\name{est}
\alias{est}
\title{Estimate a model using MML}
\usage{
est(
  dataSrc,
  predicate = NULL,
  group = NULL,
  model = c("1PL", "2PL"),
  fixed_param = NULL,
  se = TRUE,
  priorA = c("none", "lognormal", "normal"),
  priorA_mu = ifelse(priorA == "lognormal", 0, 1),
  priorA_sigma = ifelse(priorA == "lognormal", 0.5, 0.2)
)
}
\arguments{
\item{dataSrc}{a matrix, long format data.frame or a dexter database}

\item{predicate}{logical predicate to filter dataSrc, has no effect when dataSrc is a matrix}

\item{group}{if dataSrc is a matrix then a vector of length nrows, otherwise one or more person
properties together grouping people. See details.}

\item{model}{1PL or 2PL, see details.}

\item{fixed_param}{data.frame with columns: item_id, item_score, beta and, if model is 2PL, also alpha.}

\item{se}{should standard errors be determined. For large datasets with many items this can take some time. Set
to false to save time.}

\item{priorA}{if the estimation does not converge or gives extreme results, usually in an adaptive test or with too few
observations for some items in your data, you can attempt to use a prior to improve the results. Choice of
lognormal or normal}

\item{priorA_mu}{first moment of prior distribution on discrimination parameters.}

\item{priorA_sigma}{second moment of prior distribution on discrimination parameters.}
}
\value{
a list of things, to do: organize return value
}
\description{
Estimate a one or two parameter model using Marginal Maximum Likelihood
}
\details{
In a 1PL item difficulties are estimated according to the Nominal Response Model.
In a 2PL, the items also get a discrimination. This
can be interpreted as a noise factor in the item measurement analogous to item test
correlations in classical test theory. Both the 1PL and 2PL model can handle polytomous data and respect the item
scores. Missing categories (e.g. an item scored 0,1,4) are allowed.

Specifying grouping variables for test takers is very important in MML estimation. Failure to include
relevant grouping can seriously bias parameter and subsequent ability estimation.
}
\section{choice of calibration method}{


Note that MML estimation requires extra assumptions about the population distribution compared to CML.
Consequently there is rarely a good reason to use MML estimation for an 1PL since it is an exponential family
model and can be better estimated with CML. Only in case of adaptive data (where CML is not allowed) should you
use MML to estimate a 1PL.

A 2PL cannnot be estimated using CML, except in the case of complete data (see the interaction model in dexter).
So for 2PL models MML is usually the method of choice.
}

\section{troubleshooting}{


The EM algorithm tries to converge on a solution up to a precision of 0.0001. It usually succeeds.
If it is not successful
a message or warning is given (dependent on the severity of the situation). A message can usually be ignored if
you are happy with a slightly lower precision. In case of a warning, the (less precise) results are
still returned to facilitate identification of the problem but you should generally not trust the results very much.

The following possible solutions can be tried in such a case:

\describe{
\item{omit fixed parameters}{If you use fixed parameters, try to calibrate without fixed parameters
first and plot the results against your fixed parameters. If these do not fall approximately on
a straight line, you might need to omit some of the fixed parameters that show the most misfit.}
\item{linear subsets 1PL}{If your testdata includes adaptive or random tests, but it contains
a significant subset of items that was administered only as linear or multi stage tests, you can use the functions fit_enorm
or fit_enorm_mst in dexter and dexterMST respectively to fit a CML solution on this subset. Next you
can fit the complete dataset with dexterMML while fixing the parameters of the linearly administered items.}
\item{priors in 2PL}{If the results of the calibration are very extreme (e.g. parameters with absolute values >60) it
might be necessary to use a prior distribution on the discrimination parameters.
This may happen with adaptive test data.}
}
}

