% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values.R
\name{plausible_values.mml}
\alias{plausible_values.mml}
\title{plausible values}
\usage{
plausible_values.mml(
  dataSrc,
  parms,
  predicate = NULL,
  covariates = NULL,
  npv = 1
)
}
\arguments{
\item{dataSrc}{a connection to a dexter database, a matrix, or a data.frame with columns: person_id, item_id, item_score}

\item{parms}{object produced by function est}

\item{predicate}{An optional expression to subset data, if NULL all data is used}

\item{covariates}{character vector indicating discrete person properties or, in case dataSrc is a matrix,
a vector of groups (integer or character)}

\item{npv}{number of plausible values to draw per person}
}
\value{
data.frame
}
\description{
This function uses the SVE algorithm (Marsman,...;) to produce a dependent sample of plausible values
}
