% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_est.R
\name{plot.parms_mml}
\alias{plot.parms_mml}
\title{Plot for fitted MML models}
\usage{
\method{plot}{parms_mml}(x, items = NULL, nbins = 5, ci = 0.95, ...)
}
\arguments{
\item{x}{object produced by fit_1pl or fit_2pl}

\item{items}{item_id's of items to plot, if NULL, one plot for each item is made}

\item{nbins}{number of ability groups}

\item{ci}{confidence interval for the error bars, between 0 and 1. 
Default = 0.95 for a 95\% confidence interval}

\item{...}{further arguments to plot}
}
\description{
The plot shows 'fit' by comparing the expected score based on the model (grey line)
with the average scores based on the data (black line with dots) for groups of students
with similar estimated ability.
}
