% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values.R
\name{sim_2pl}
\alias{sim_2pl}
\title{Simulate data for a 2pl}
\usage{
sim_2pl(pars, theta)
}
\arguments{
\item{pars}{data.frame with columns alpha, item_score and beta}

\item{theta}{vector of abilities}
}
\value{
matrix with persons as rows and items as columns
}
\description{
note: to simulate form a 1pl, you can omit the alpha column or set it to 1
}
