% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ability.R
\name{ability.mml}
\alias{ability.mml}
\title{Abilities for a 1 and 2pl}
\usage{
ability.mml(
  dataSrc,
  parms,
  predicate = NULL,
  method = c("MLE", "WLE"),
  unweight = FALSE
)
}
\arguments{
\item{dataSrc}{a connection to a dexter database, a matrix, or a data.frame with columns: person_id, item_id, item_score}

\item{parms}{object produced by function fit_1pl or fit_2pl or possibly a data.frame of parameters}

\item{predicate}{An optional expression to subset data, if NULL all data is used}

\item{method}{Maximum Likelihood (MLE), Weighted Likelihood (WLE)}

\item{unweight}{(experimental) whether to use the weighted score or the unweighted score. Has no effect for 1PL, see details}
}
\value{
data.frame with variables person_id and theta and se
}
\description{
Abilities for a 1 and 2pl
}
\details{
for a 2pl you have the option to use the weighted or unweighted score. The weighted score gives ML estimates
for the regular 2pl. The unweighted score is an experimental adaptation where ability is computed conditional on the unweighted
sumscore of the respondent. This means people with the same unweighted score get the same ability estimate. This is 
option is experimental and awaiting future adaptations and checks.

When using a data.frame of parameters, be sure that you use the correct parametrisation. See \code{\link{fit_1pl}} for details.
}
