% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_est.R
\name{coef.parms_mml}
\alias{coef.parms_mml}
\title{Extract information from MML fit object}
\usage{
\method{coef}{parms_mml}(object, what = c("items", "populations", "likelihood"))
}
\arguments{
\item{object}{object returned by est}

\item{what}{information to extract}
}
\description{
Extract information from MML fit object
}
