% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_est.R
\name{est}
\alias{est}
\title{Estimate a model using MML}
\usage{
est(
  dataSrc,
  predicate = NULL,
  group = NULL,
  model = c("1PL", "2PL"),
  se = TRUE,
  priorA = FALSE,
  priorA_mu = 0,
  priorA_sigma = 0.5
)
}
\arguments{
\item{dataSrc}{a matrix, long format data.frame or a dexter database}

\item{predicate}{logical predicate to filter dataSrc, has no effect when dataSrc is a matrix}

\item{group}{if dataSrc is a matrix then a vector of length nrows, otherwise one or more person
properties together grouping people. See details.}

\item{model}{1PL or 2PL, see details.}

\item{se}{should standard errors be determined. For large datasets with many items this can take some time. Set
to false to save time.}

\item{priorA}{set to TRUE to enable a lognormal prior distribution on the discrimination parameters. This is
a good idea if you have data from an adaptive test and/or the algorithm fails to converge}

\item{priorA_mu}{first moment of prior distribution on discrimination parameters. 0 is a sensible defa}
}
\value{
a list of things, to do: organize return value
}
\description{
Estimate a one or two parameter model using Marginal Maximum Likelihood
}
\details{
In a 1PL item difficulties on a logistic scale are estimated in a marginal Nominal Response Model.
In a 2PL model items also get a discrimination. This
can be interpreted as a noise factor in the item measurement analogous to item test
correlations in classical test theory. Both the 1PL and 2PL model can handle polytomous data (to do: 2PL poly not finished yet) and respect the item
scores. Missing categories (e.g. an item scored 0,1,4) are allowed.

Specifying grouping variables for test takers is very important in MML estimation. Failure to include
relevant grouping can seriously bias parameter and subsequent ability estimation.

Note that MML estimation requires extra assumptions about the population distribution compared to CML.
Consequently there is rarely a good reason to use MML estimation for an 1PL since it is an exponential family
model and can be better estimated with CML. Only in case of adaptive data (where CML is not allowed) should you
use MML to estimate a 1PL.

A 2PL cannnot be estimated using CML, except in case of complete data (see the interaction model in dexter).
So for 2PL models MML is usually the method of choice.
}
