% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values.R
\name{plausible_values.mml}
\alias{plausible_values.mml}
\title{plausible values}
\usage{
plausible_values.mml(
  dataSrc,
  parms,
  predicate = NULL,
  covariates = NULL,
  npv = 1
)
}
\arguments{
\item{dataSrc}{a connection to a dexter database, a matrix, or a data.frame with columns: person_id, item_id, item_score}

\item{parms}{object produced by function fit_1pl or fit_2pl or possibly a data.frame of parameters}

\item{predicate}{An optional expression to subset data, if NULL all data is used}

\item{covariates}{character vector indicating discrete person properties or, in case dataSrc is a matrix,
a vector of groups (integer or character)}

\item{npv}{number of plausible values to draw per person}
}
\value{
data.frame
}
\description{
This function uses the Single Variable Exchange algorithm
(Murray, Ghahramani & MacKay, 2012; see Marsman, Maris, Bechger & Glas, 2014, for application to IRT models)
to produce a dependent sample of plausible values.
}
\details{
This function uses the Single Variable Exchange algorithm
(Murray, Ghahramani & MacKay,2012; see Marsman, Maris, Bechger & Glas, 2014, for application to IRT models)
to produce a dependent sample of plausible values. The settings are such that the autocorrelation is reasonably low
for tests with up to 40 items. To further lower the autocorrelation, e.g. for longer tests, the user can
simply draw more plausible values than needed (see argument npv) and apply additional thinning.

When using a data.frame of parameters, be sure that you use the correct parametrisation. See \code{link{fit_1pl}} for details.
}
\references{
Marsman, M., Maris, G., Bechger, T. M., and Glas, C.A.C. (2014) Composition algorithms for Conditional Distributions.
PhD thesis

Murray, I., Ghahramani, Z., & MacKay, D. (2012). MCMC for doubly-intractable distributions. via: https://arxiv.org/abs/1206.6848
}
